NEST/backend/main.py
from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
from typing import Optional, List
import uvicorn
from datetime import datetime, timedelta
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

app = FastAPI(
    title="NEST API",
    description="Mental Health Support Platform API",
    version="1.0.0"
)

# CORS middleware configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # In production, replace with specific origins
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Security
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# Models
class UserBase(BaseModel):
    username: str
    email: Optional[str] = None

class UserCreate(UserBase):
    password: str

class User(UserBase):
    id: int
    is_active: bool
    created_at: datetime

    class Config:
        from_attributes = True

class Token(BaseModel):
    access_token: str
    token_type: str

class TokenData(BaseModel):
    username: Optional[str] = None

# Routes
@app.get("/")
async def root():
    return {"message": "Welcome to NEST API"}

@app.post("/users/", response_model=User)
async def create_user(user: UserCreate):
    # TODO: Implement user creation with password hashing
    return {"message": "User creation endpoint"}

@app.post("/token", response_model=Token)
async def login_for_access_token():
    # TODO: Implement token generation
    return {"message": "Token generation endpoint"}

@app.get("/users/me/", response_model=User)
async def read_users_me():
    # TODO: Implement user profile retrieval
    return {"message": "User profile endpoint"}

# Emotion Analysis endpoints
@app.post("/analyze/emotion")
async def analyze_emotion(text: str):
    # TODO: Implement emotion analysis using transformer models
    return {"message": "Emotion analysis endpoint"}

# AI Chat endpoints
@app.post("/chat")
async def chat_with_ai(message: str):
    # TODO: Implement AI chat functionality
    return {"message": "AI chat endpoint"}

# Meditation and Exercises endpoints
@app.get("/exercises/meditation")
async def get_meditation_exercises():
    # TODO: Implement meditation exercises retrieval
    return {"message": "Meditation exercises endpoint"}

@app.get("/exercises/breathing")
async def get_breathing_exercises():
    # TODO: Implement breathing exercises retrieval
    return {"message": "Breathing exercises endpoint"}

# Emotion Tracking endpoints
@app.post("/track/emotion")
async def track_emotion():
    # TODO: Implement emotion tracking
    return {"message": "Emotion tracking endpoint"}

@app.get("/reports/emotion")
async def get_emotion_report():
    # TODO: Implement emotion report generation
    return {"message": "Emotion report endpoint"}

if __name__ == "__main__":
    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True) 