NEST/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
pika==1.3.2
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.5
aiofiles==23.2.1
python-magic==0.4.27 